# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Water Tank Monitoring API
  description: Backend REST API & Realtime Monitoring Specification for Integrated Hospital Water Tank Monitoring Research Project (RSA UGM).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
...
  /api/tanks:
    get:
      summary: List all water tanks
      description: Returns all monitored water tanks with their latest reading, active alert, and attached sensor nodes.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TankDetail'
    post:
      summary: Create new tank (Admin only)
      description: Registers a new water tank into the monitoring system.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTankInput'
      responses:
        '201':
          description: Tank created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Tank'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
  /api/tanks/{id}:
    get:
      summary: Get tank details
      description: Returns single tank details along with the 50 most recent readings and 10 recent alerts.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TankWithRecentHistory'
        '404':
          description: Tank not found
    patch:
      summary: Update tank metadata / threshold limits (Admin only)
      description: Update specific fields of a water tank.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTankInput'
      responses:
        '200':
          description: Tank updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Tank'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: Tank not found
    put:
      summary: Replace / update tank data (Admin only)
      description: Idempotent update for water tank properties.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTankInput'
      responses:
        '200':
          description: Tank updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Tank'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: Tank not found
    delete:
      summary: Delete tank (Admin only)
      description: Remove a tank master record and log audit trail.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tank deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tank deleted successfully"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: Tank not found
  /api/tanks/{id}/readings:
    get:
      summary: Get historical time-series readings for a tank
      description: Fetch time-series water level readings for a specific tank for charting and trend analysis.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: hours
          in: query
          required: false
          schema:
            type: integer
            default: 24
          description: Number of historical hours to query
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 500
          description: Maximum number of records to return
      responses:
        '200':
          description: List of time-series readings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tankId:
                    type: string
                    format: uuid
                  hours:
                    type: integer
                    example: 24
                  count:
                    type: integer
                    example: 288
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reading'
...
components:
  schemas:
...
    Tank:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Tangki Atap Gedung Utama"
        location:
          type: string
          nullable: true
          example: "Atap Gedung A (Utama), Lantai 5"
        capacityLiters:
          type: string
          example: "10000.00"
        minThresholdPercent:
          type: string
          example: "30.00"
        maxThresholdPercent:
          type: string
          example: "90.00"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTankInput:
      type: object
      required:
        - name
        - capacityLiters
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Tangki Gedung Laboratorium"
        location:
          type: string
          maxLength: 150
          example: "Gedung Lab Terpadu Lantai 4"
        capacityLiters:
          type: number
          example: 6000
        minThresholdPercent:
          type: number
          default: 30
          example: 30
        maxThresholdPercent:
          type: number
          default: 90
          example: 90
    UpdateTankInput:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacityLiters:
          type: number
        minThresholdPercent:
          type: number
        maxThresholdPercent:
          type: number
    Reading:
      type: object
      properties:
        id:
          type: integer
          example: 14205
        tankId:
          type: string
          format: uuid
        sensorNodeId:
          type: string
          format: uuid
          nullable: true
        levelPercent:
          type: string
          example: "72.40"
        volumeLiters:
          type: string
          example: "7240.00"
        rawValue:
          type: string
          nullable: true
          example: "812.00"
        rssi:
          type: integer
          nullable: true
          example: -78
        recordedAt:
          type: string
          format: date-time
...
    Alert:
      type: object
      properties:
        id:
          type: integer
          example: 89
        tankId:
          type: string
          format: uuid
        tankName:
          type: string
          example: "Tangki Gedung IBS"
        tankLocation:
          type: string
          nullable: true
        level:
          type: string
          enum: [WARNING, CRITICAL]
          example: "CRITICAL"
        message:
          type: string
          example: "Level air tangki IBS CRITICAL: 24.5%"
        status:
          type: string
          enum: [active, resolved]
          example: "active"
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
...
    SensorNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tankId:
          type: string
          format: uuid
        gatewayId:
          type: string
          format: uuid
          nullable: true
        deviceCode:
          type: string
          example: "STM32-GEDUNG-A"
        sensorType:
          type: string
          example: "ultrasonic"
        status:
          type: string
          example: "active"
        lastRssi:
          type: integer
          nullable: true
          example: -78
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TankDetail:
      allOf:
        - $ref: '#/components/schemas/Tank'
        - type: object
          properties:
            latestReading:
              $ref: '#/components/schemas/Reading'
              nullable: true
            activeAlert:
              $ref: '#/components/schemas/Alert'
              nullable: true
            nodes:
              type: array
              items:
                $ref: '#/components/schemas/SensorNode'
    TankWithRecentHistory:
      allOf:
        - $ref: '#/components/schemas/Tank'
        - type: object
          properties:
            recentReadings:
              type: array
              items:
                $ref: '#/components/schemas/Reading'
            recentAlerts:
              type: array
              items:
                $ref: '#/components/schemas/Alert'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token obtained from /api/auth/login
    GatewayKeyAuth:
      type: apiKey
      in: header
      name: X-Gateway-Key
      description: Pre-shared API Key header for LoRa Gateway synchronization